+++ client/httplib/httplib_test.go
@@ -433,3 +433,7 @@ func TestBeegoHTTPRequestXMLBody(t *testing.T) {
 	assert.Nil(t, err)
 	assert.NotNil(t, req.req.GetBody)
 }
+
+// TODO
+func TestBeegoHTTPRequest_ResponseForValue(t *testing.T) {
+}